修复点击备注以后数量都默认为1的问题。

+++ CanDao.Pos.UI.Library/View/SetDishTasteAndDietWindow.xaml.cs
@@ -42,7 +42,6 @@ namespace CanDao.Pos.UI.Library.View
 
         public int DishNum
         {
-            get { return ((SetDishTasteAndDietWndVm)DataContext).DishNum; }
-        }
+            get { return (int)((SetDishTasteAndDietWndVm)DataContext).DishNum; }}
     }
 }
